[chore]: refactor out base classes for tree and tree-item
