+++ packages/web-components/src/tree/index.ts
@@ -1,3 +1,4 @@
+export { BaseTree } from './tree.base.js';
 export { Tree } from './tree.js';
 export { template as TreeTemplate } from './tree.template.js';
 export { styles as TreeStyles } from './tree.styles.js';
